Add a red dot next to nametag
Add a red dot next to nametag to draw user attention upon loading the page. This would prompt the user to hover over or click the dot/nametag, which would bring up useful contact information in a future modal popup feature.

A small hover animation is also added to the dot for visual appeal.