Add basic modal box to show more info upon click
Upon clicking the nametag on the landing page, a basic modal popup will be displayed, showing placeholder text for now.

The modal style will be modified for visual appeal in the future, but will remain consistent for all modals on this site.

The modal content will be defined semantically in HTML on a per basis.